requirements.txt
pyboy>=2.0.0
aiohttp
ollama
pillow
